notification model+tests and small improvements on reminder as part of application reminder
